+++ agent-samples/xr-render-demo/yaml/xr_media_hub.yaml
@@ -31,6 +31,3 @@ enable_token_server: true
 # Enable and set recordings_dir in video_mcp_server.yaml to unlock historical tools.
 video_recording:
   enabled: false
-
-# GPU index for cloudxr + render-mcp (LOVR inherits from render-mcp).
-gpu_index: 0
